refactor: 빌드 에러 수정 및 코드 리뷰 반영
