+++ src/RoomSearch/index.ts
@@ -1,0 +1,3 @@
+export { default as SearchBar } from '@/RoomSearch/components/SearchBar';
+export { default as ResultList } from '@/RoomSearch/components/ResultList';
+export { default as Selection } from '@/RoomSearch/components/Selection';
