--- src/Search/index.ts
@@ -1,3 +1,0 @@
-export { default as SearchBar } from '@/Search/components/SearchBar';
-export { default as ResultList } from '@/Search/components/ResultList';
-export { default as Selection } from '@/Search/components/Selection';
